Change "show more repos" to simply extend the repo list
Gitea's index page features a repository list, which by default only shows
at most 15 repos. If there are more, a "show more repos" link is
displayed at the bottom.

Previously this was a link to the user's profile, but that is not helpful,
as not all repos (e.g. those owned by organizations) are shown there.